requirements.txt
# Framework principal
Django==5.0.4

# API REST
djangorestframework==3.15.1

# Autenticación OAuth2
django-oauth-toolkit==2.4.0

# Base de datos PostgreSQL
psycopg2-binary==2.9.9

# Manipulación de datos y CSV
pandas==2.2.2

# Timezone
pytz==2024.1

# Servidor WSGI
gunicorn==22.0.0